requirements.txt
numpy
pandas
python-mnist
torch
torchvision
